+++ .github/workflows/before_merge.yaml
@@ -27,10 +27,17 @@ jobs:
           python3 -m pip install --upgrade pip                    
           python3 -m pip install -r requirements-dev.txt
           python3 -m pip install pytest-github-report
-      - name: Run Tests
+      - name: Run Lowering Tests
         env:
           pytest_verbosity: 2
-          pytest_report_title: "⭐️ Pytest Results ⭐️"
+          pytest_report_title: "⭐️ Aten → TTNN Lowering Tests"
         run: | 
           source venv/bin/activate
-          python3 -m pytest --github-report tests/*.py -s
+          python3 -m pytest --github-report tests/lowering/ -s
+      - name: Run Model Tests
+        env:
+          pytest_verbosity: 2
+          pytest_report_title: "⭐️ Model Tests"
+        run: | 
+          source venv/bin/activate
+          python3 -m pytest --github-report tests/models/ -s
