+++ tests/lowering/tensor_manipulation/test_expand.py
