+++ torch_ttnn/passes/lowering/permute_reshape_tuple.py
